Consider this as INITAL COMMIT
The src folder contains the pre-processing files. The first file was created for generating all types of actions. Then a second file (v2) was created to generate more focused data for our analysis.

After this commit, we will remove the first file, and create only a single pre-processing file as github will control the versioning by default.

NOTE: the actual data will not be shared due to permission.