partially developed crystal island env simulator
need to test the simulator. right now it should be able to step through the environment, generate next states when given actions. might need to flesh out further.

also fixed the pre-processor for creating states with dedicated positional features that map between student_action and feature position. for example, feature position 5 is s_end_game, and action 5 is also s_end_game.

if we change the features/actions, we might need to fix this as well as the environment functions.