GAIL implemented
GAIL is working. Might need refactoring.

Refactored Crystal Island.

The current result looks optimistic when testing with random and test sets. But if we split data into high-low nlg, D performance for complement data is the same as the test set, meaning there is no significant difference that D knows.